Floatify Coord and more debugging information

+++ include/AxisControlClass.h
@@ -9,8 +9,8 @@ class AxisControlClass {
         void moveZ(int length, int speed);
         void moveE(int length, int speed);
         void moveAxisInf(int speed);
-        static void moveTest(int tx, int ty);
-        static void calculateMovements(int target_x, int target_y, int* stp_x, int* stp_y);
+        static void moveTest(float tx, float ty);
+        static void calculateMovements(float target_x, float target_y, int* stp_x, int* stp_y);
         static void moveTestY(int speed, int steps);
         static void moveTestX(int speed, int steps);
 };
